+++ pybricksdev/connections/__init__.py
@@ -1,0 +1,2 @@
+# SPDX-License-Identifier: MIT
+# Copyright (c) 2022 The Pybricks Authors